requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.1
django-ckeditor==6.5.1
django-cleanup==6.0.0
django-js-asset==2.0.0
django-resized==1.0.2
frozenlist==1.3.1
gunicorn==20.1.0
idna==3.4
multidict==6.0.2
mysqlclient==2.1.1
Pillow==9.2.0
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.4
urllib3==1.26.12
whitenoise==6.2.0
yarl==1.8.1
